requirements.txt
Django==1.9.7
django-generic-scaffold==0.3.1
six==1.10.0
